+++ oqs/src/kem.rs
@@ -3,9 +3,13 @@
 //! See [`Kem`] for the main functionality.
 //! [`Algorithm`] lists the available algorithms.
 use alloc::vec::Vec;
+use rand_core::CryptoRngCore;
 
 use core::ptr::NonNull;
 use core::str::FromStr;
+
+use ::kem::Decapsulate as RustCryptoDecapsulate;
+use ::kem::Encapsulate as RustCryptoEncapsulate;
 
 #[cfg(not(feature = "std"))]
 use cstr_core::CStr;
@@ -180,9 +184,9 @@ implement_kems! {
     ("classic_mceliece") ClassicMcEliece6960119f: OQS_KEM_alg_classic_mceliece_6960119f,
     ("classic_mceliece") ClassicMcEliece8192128: OQS_KEM_alg_classic_mceliece_8192128,
     ("classic_mceliece") ClassicMcEliece8192128f: OQS_KEM_alg_classic_mceliece_8192128f,
-    ("hqc") Hqc128: OQS_KEM_alg_hqc_1,
-    ("hqc") Hqc192: OQS_KEM_alg_hqc_3,
-    ("hqc") Hqc256: OQS_KEM_alg_hqc_5,
+    ("hqc") Hqc1: OQS_KEM_alg_hqc_1,
+    ("hqc") Hqc3: OQS_KEM_alg_hqc_3,
+    ("hqc") Hqc5: OQS_KEM_alg_hqc_5,
     ("kyber") Kyber512: OQS_KEM_alg_kyber_512,
     ("kyber") Kyber768: OQS_KEM_alg_kyber_768,
     ("kyber") Kyber1024: OQS_KEM_alg_kyber_1024,
@@ -498,5 +502,87 @@ impl Kem {
         // this is safe to do, as we have initialised them now.
         unsafe { ss.bytes.set_len(kem.length_shared_secret) };
         Ok(ss)
+    }
+}
+
+/// Encapsulator uses the public key of a KEM to encapsulate keys. This struct implements RustCrypto's Encapsulate trait.
+///
+/// # Example
+/// ```rust
+/// # if !cfg!(feature = "ml_kem") { return; }
+/// use oqs;
+/// use kem::{Encapsulate,Decapsulate};
+/// use rand_core::OsRng;
+/// oqs::init();
+/// let kem = oqs::kem::Kem::new(oqs::kem::Algorithm::MlKem512).unwrap();
+/// let (pk, sk) = kem.keypair().unwrap();
+/// let encap = oqs::kem::Encapsulator::new(&kem, pk);
+/// let decap = oqs::kem::Decapsulator::new(&kem, sk);
+/// // Note that encapsulate ignores the provided rng
+/// let (ct, ss) = encap.encapsulate(&mut OsRng).unwrap();
+/// let ss2 = decap.decapsulate(&ct).unwrap();
+/// assert_eq!(ss, ss2);
+/// ```
+pub struct Encapsulator<'a> {
+    scheme: &'a Kem,
+    pk: PublicKey,
+}
+
+impl<'a> Encapsulator<'a> {
+    /// Creates a new [`Encapsulator`].
+    pub fn new(scheme: &'a Kem, pk: PublicKey) -> Self {
+        Encapsulator { scheme, pk }
+    }
+}
+
+/// Decapsulator struct containing a kem scheme and a secret key.
+///
+/// # Example
+/// ```rust
+/// # if !cfg!(feature = "ml_kem") { return; }
+/// use oqs;
+/// use kem::{Encapsulate,Decapsulate};
+/// use rand_core::OsRng;
+/// oqs::init();
+/// let kem = oqs::kem::Kem::new(oqs::kem::Algorithm::MlKem512).unwrap();
+/// let (pk, sk) = kem.keypair().unwrap();
+/// let encap = oqs::kem::Encapsulator::new(&kem, pk);
+/// let decap = oqs::kem::Decapsulator::new(&kem, sk);
+/// // Note that encapsulate ignores the provided rng
+/// let (ct, ss) = encap.encapsulate(&mut OsRng).unwrap();
+/// let ss2 = decap.decapsulate(&ct).unwrap();
+/// assert_eq!(ss, ss2);
+/// ```
+pub struct Decapsulator<'a> {
+    scheme: &'a Kem,
+    sk: SecretKey,
+}
+
+impl<'a> Decapsulator<'a> {
+    /// Creates a new [`Decapsulator`].
+    pub fn new(scheme: &'a Kem, sk: SecretKey) -> Self {
+        Decapsulator { scheme, sk }
+    }
+}
+
+impl<'a> RustCryptoEncapsulate<Ciphertext, SharedSecret> for Encapsulator<'a> {
+    type Error = crate::Error;
+
+    fn encapsulate(
+        &self,
+        _csprng: &mut impl CryptoRngCore,
+    ) -> core::result::Result<(Ciphertext, SharedSecret), Self::Error> {
+        self.scheme.encapsulate(&self.pk)
+    }
+}
+
+impl<'a> RustCryptoDecapsulate<Ciphertext, SharedSecret> for Decapsulator<'a> {
+    type Error = crate::Error;
+
+    fn decapsulate(
+        &self,
+        encapsulated_key: &Ciphertext,
+    ) -> core::result::Result<SharedSecret, Self::Error> {
+        self.scheme.decapsulate(&self.sk, encapsulated_key)
     }
 }